+++ psutil/_pslinux.py
@@ -704,6 +704,19 @@ if os.path.exists("/sys/devices/system/cpu/cpufreq") or \
             ret.append(_common.scpufreq(curr, min_, max_))
         return ret
 
+elif os.path.exists("/proc/cpuinfo"):
+    def cpu_freq():
+        """Alternate implementation using /proc/cpuinfo.
+        min and max frequencies are not available and are set to None.
+        """
+        ret = []
+        with open_binary('%s/cpuinfo' % get_procfs_path()) as f:
+            for line in f:
+                if line.lower().startswith(b'cpu mhz'):
+                    key, value = line.split(b'\t:', 1)
+                    ret.append(_common.scpufreq(float(value), None, None))
+        return ret
+
 
 # =====================================================================
 # --- network